pyproject.toml
[project]
name = 'demos'
version = '0.0'
authors = [{ name = 'Jordan Cook' }]
requires-python = '>=3.10'
dependencies = [
    'attrs',
    'ipdb',
    'ipython',
    'datasette',
    'pandas',
    'pillow',
    'platformdirs',
    'pyarrow',
    'python-dateutil',
    'pytimeparse',
    'pyyaml',
    'requests',
    'requests-cache',
    'rich',
    'tablib',
    'ujson',
    'whenever',
    'time_machine',
]

[dependency-groups]
dev = [
    'nox',
    'pre-commit',
    'pytest',
    'requests-mock',
]

[tool.ruff]
fix = true
unsafe-fixes = true
line-length = 100
output-format = 'grouped'

[tool.ruff.format]
quote-style = 'single'

[tool.ruff.lint]
select = ['B', 'C4', 'C90', 'E', 'F', 'I']

[tool.ruff.lint.pycodestyle]
max-line-length = 120
